refactor ShardingRewriteEngine and EncryptRewriteEngine
